+++ etna/core/__init__.py
@@ -2,3 +2,4 @@ from etna.core.mixins import BaseMixin
 from etna.core.mixins import SaveMixin
 from etna.core.mixins import StringEnumWithRepr
 from etna.core.saving import AbstractSaveable
+from etna.core.utils import load