+++ etna/pipeline/mixins.py
@@ -1,10 +1,17 @@
+import pathlib
+import tempfile
+import zipfile
 from copy import deepcopy
+from typing import Any
+from typing import Optional
 from typing import Sequence
 
 import numpy as np
 import pandas as pd
 from typing_extensions import get_args
 
+from etna.core import SaveMixin
+from etna.core import load
 from etna.datasets import TSDataset
 from etna.models import ModelType
 from etna.models import NonPredictionIntervalContextIgnorantAbstractModel
@@ -80,3 +87,106 @@ class ModelPipelinePredictMixin:
         else:
             raise NotImplementedError(f"Unknown model type: {self.model.__class__.__name__}!")
         return results
+
+
+class SaveModelPipelineMixin(SaveMixin):
+    """Implementation of ``AbstractSaveable`` abstract class for pipelines with model inside.
+
+    It saves object to the zip archive with 2 files:
+
+    * metadata.json: contains library version and class name.
+
+    * object.pkl: pickled without model, transforms and ts.
+
+    * model.zip: saved model.
+
+    * transforms: folder with saved transforms.
+    """
+
+    model: ModelType
+    transforms: Sequence[Transform]
+    ts: Optional[TSDataset]
+
+    def save(self, path: pathlib.Path):
+        """Save the object.
+
+        Parameters
+        ----------
+        path:
+            Path to save object to.
+        """
+        model = self.model
+        transforms = self.transforms
+        ts = self.ts
+        try:
+            # extract attributes we can't easily save
+            delattr(self, "model")
+            delattr(self, "transforms")
+            delattr(self, "ts")
+
+            # save the remaining part
+            super().save(path=path)
+        finally:
+            self.model = model
+            self.transforms = transforms
+            self.ts = ts
+
+        with zipfile.ZipFile(path, "a") as archive:
+            with tempfile.TemporaryDirectory() as _temp_dir:
+                temp_dir = pathlib.Path(_temp_dir)
+
+                # save model separately
+                model_save_path = temp_dir / "model.zip"
+                model.save(model_save_path)
+                archive.write(model_save_path, "model.zip")
+
+                # save transforms separately
+                transforms_dir = temp_dir / "transforms"
+                transforms_dir.mkdir()
+                num_digits = len(str(len(transforms) - 1))
+                for i, transform in enumerate(transforms):
+                    save_name = f"{i:0{num_digits}d}.zip"
+                    transform_save_path = transforms_dir / save_name
+                    transform.save(transform_save_path)
+                    archive.write(transform_save_path, f"transforms/{save_name}")
+
+    @classmethod
+    def load(cls, path: pathlib.Path, ts: Optional[TSDataset] = None) -> Any:
+        """Load an object.
+
+        Parameters
+        ----------
+        path:
+            Path to load object from.
+        ts:
+            TSDataset to set into loaded pipeline.
+
+        Returns
+        -------
+        :
+            Loaded object.
+        """
+        obj = super().load(path=path)
+        obj.ts = deepcopy(ts)
+
+        with zipfile.ZipFile(path, "r") as archive:
+            with tempfile.TemporaryDirectory() as _temp_dir:
+                temp_dir = pathlib.Path(_temp_dir)
+
+                archive.extractall(temp_dir)
+
+                # load model
+                model_path = temp_dir / "model.zip"
+                obj.model = load(model_path)
+
+                # load transforms
+                transforms_dir = temp_dir / "transforms"
+                transforms = []
+
+                if transforms_dir.exists():
+                    for path in sorted(transforms_dir.iterdir()):
+                        transforms.append(load(path))
+
+                obj.transforms = transforms
+
+        return obj